+++ src/app/view/mission/PaginatedMissionListComponent.js
@@ -59,7 +59,8 @@ class PaginatedMissionListComponent extends Component {
 					this.props.admin || this.props.user ? state.currentFilters.status || "all" : null,
 					this.props.width === "xs" || state.currentFilters.editor,
 					state.currentFilters.complete,
-					this.props.user && !this.props.admin ? this.props.user.id : null
+					this.props.user && !this.props.admin ? this.props.user.id : null,
+					state.currentFilters.sort
 				)
 				.then(missions => {
 					if(currentToken === this.apiCallToken) {
@@ -80,7 +81,8 @@ class PaginatedMissionListComponent extends Component {
 				this.props.admin || this.props.user  ? state.currentFilters.status || "all" : null,
 				this.props.width === "xs" || state.currentFilters.editor,
 				state.currentFilters.complete,
-				this.props.user && !this.props.admin ? this.props.user.id : null
+				this.props.user && !this.props.admin ? this.props.user.id : null,
+				state.currentFilters.sort
 			)
 			.then(missions => {
 				if(currentToken === this.apiCallToken) {
@@ -193,7 +195,7 @@ class PaginatedMissionListComponent extends Component {
 			<Grid container spacing={16}>
 				<Grid item hidden={{only: "xs"}} sm={4} md={3} lg={2}>
 					<Typography variant="subheading">{I18n.t("Filters")}</Typography>
-					<MissionsFilters status={this.props.admin || this.props.user} completeness={true} values={this.state.currentFilters} onChange={d => this.setState({ currentFilters: d })} />
+					<MissionsFilters status={this.props.admin || this.props.user} completeness={true} values={this.state.currentFilters} onChange={d => this.setState({ currentFilters: d })} sorting={this.state.tab === 0} />
 				</Grid>
 				<Grid item xs={12} hidden={{smUp: true}}>
 					<ExpansionPanel>
@@ -201,7 +203,7 @@ class PaginatedMissionListComponent extends Component {
 							<Typography variant="subheading">{I18n.t("Filters")}</Typography>
 						</ExpansionPanelSummary>
 						<ExpansionPanelDetails>
-							<MissionsFilters values={this.state.currentFilters} onChange={d => this.setState({ currentFilters: d })} />
+							<MissionsFilters values={this.state.currentFilters} sorting={this.state.tab === 0} onChange={d => this.setState({ currentFilters: d })} />
 						</ExpansionPanelDetails>
 					</ExpansionPanel>
 				</Grid>