+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Translations of automatic detections feature types in new mission page
 - New missions based on Osmose or automatic detections have a thumbnail associated in mission list
 - Thumbnail of mission is shown in mission description page
+- Missions list can be sorted by pertinence, creation date and completeness
 
 
 ## 0.5.6 - 2019-04-01